Changed pBegin & pEnd type for bigger token counts
Used to be `unsigned short`, is now `unsigned int` because when an integer overflow happens. the string capturer freaks out.

+++ stringtokenizer.lua
@@ -7,7 +7,7 @@ StringTokenizer.__index = StringTokenizer
 
 ffi.cdef [[
 typedef struct {
-    unsigned short pBegin, pEnd;
+    unsigned int pBegin, pEnd;
     unsigned char captureIndex;
 } StringTokenizerToken;
 ]]
